modularising pages into components
